requirements.txt
Django>=3.0.6,<3.1
uWSGI>=2.0.18,<2.1
djangorestframework==3.13.1
Pillow==8.4
